feat: squad is deleted from DB if the player leaving/being removed is the last plyer remaining in the squad

+++ frontend/src/components/AdminPage/UpdateSquad.vue
@@ -9,11 +9,19 @@
       <div v-if="squadMembers.length > 0">
         <label><strong>Kick: </strong></label>
         <select v-model="memberId" class="select-component">
-          <option value=0></option>
-          <option v-for="(member, index) in squadMembers" :value="member.id" :key="index">{{member.username}}</option>
+          <option value="0"></option>
+          <option
+            v-for="(member, index) in squadMembers"
+            :value="member.id"
+            :key="index"
+          >
+            {{ member.username }}
+          </option>
         </select>
       </div>
-      <button class="update-btn" type="button" @click="updateSquad">Update Squad</button>
+      <button class="update-btn" type="button" @click="updateSquad">
+        Update Squad
+      </button>
     </form>
   </div>
 </template>
@@ -25,32 +33,62 @@ export default {
     squad: Object,
     gameId: Number,
   },
-  data(){
-    return{
+  data() {
+    return {
       squadMembers: [],
-      memberId: 0
+      memberId: 0,
     };
   },
   async created() {
     for (let i = 0; i < this.squad.squadMembers.length; i++) {
-      const response = await fetch(`https://hvz-experis-api.herokuapp.com${this.squad.squadMembers[i].player}`);
+      const response = await fetch(
+        `https://hvz-experis-api.herokuapp.com${this.squad.squadMembers[i].player}`
+      );
       const data = await response.json();
-      this.squadMembers.push(data)   
+      this.squadMembers.push(data);
     }
   },
   computed: {
-    ...mapState(["apiUrl", "token"])
+    ...mapState(["apiUrl", "token"]),
   },
   methods: {
-    async kickPlayer(){
-      if(this.memberId!==0){      
-        await fetch(this.apiUrl+"/game/"+this.gameId + "/squad/" + this.squad.id +"/leave/" + this.memberId ,{
-          method: "DELETE",
-          headers: {
-            "Authorization": "Bearer " + this.token,
-            "Content-Type": "application/json"
-          },
-        })
+    async kickPlayer() {
+      if (this.memberId !== 0) {
+        await fetch(
+          this.apiUrl +
+            "/game/" +
+            this.gameId +
+            "/squad/" +
+            this.squad.id +
+            "/leave/" +
+            this.memberId,
+          {
+            method: "DELETE",
+            headers: {
+              Authorization: "Bearer " + this.token,
+              "Content-Type": "application/json",
+            },
+          }
+        );
+        //Removes the squad from the database if the player being removed is the last squad member remaining.
+        const response = await fetch(
+          "https://hvz-experis-api.herokuapp.com/api/v1" +
+            `/game/${this.gameId}/squad`
+        );
+        const data = await response.json();
+        if (data[0].squadMembers.length < 1) {
+          console.log("inne i deleten");
+          await fetch(
+            this.apiUrl + "/game/" + this.gameId + "/squad/" + this.squad.id,
+            {
+              method: "DELETE",
+              headers: {
+                Authorization: "Bearer " + this.token,
+                "Content-Type": "application/json",
+              },
+            }
+          );
+        }
       }
     },
 
@@ -70,8 +108,7 @@ export default {
         }
       );
 
-      this.kickPlayer()
-
+      this.kickPlayer();
     },
   },
 };
@@ -120,7 +157,7 @@ export default {
   background-color: rgba(132, 132, 253, 0.479);
   transition: 0.3s;
 }
-.kick-btn{
+.kick-btn {
   margin-left: 10px;
   margin-top: 10px;
   width: 40px;
@@ -130,7 +167,7 @@ export default {
   background-color: rgba(211, 30, 45, 0.479);
   transition: 0.3s;
 }
-.kick-btn:hover{
+.kick-btn:hover {
   cursor: pointer;
   background-color: rgba(230, 100, 111, 0.479);
   transition: 0.3s;
@@ -138,7 +175,6 @@ export default {
 .option-component {
   border-radius: 8px;
 }
-
 
 @media screen and (max-width: 350px) {
   .update-squad-form {
